backend/reports/urls.py
"""
Report URL configuration.
"""

from django.urls import path, include
from rest_framework.routers import DefaultRouter

from reports.views import ReportViewSet

router = DefaultRouter()
router.register(r"reports", ReportViewSet, basename="report")

urlpatterns = [
    path("", include(router.urls)),
]
