backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.3",
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.77.0",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "scikit-learn>=1.6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.10",
    "vulture>=2.16",
]
test = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-asyncio>=0.25.0",
    "pytest-mock>=3.14.0",
    "freezegun>=1.5.0",
]
